scryfall get done on server
