+++ src/app/styles/card/mtg-card.css
@@ -99,7 +99,6 @@
 
 	gap: 0.1in;
 	.mtg-card {
-		/*height: calc((var(--card-height) - 0.1in) / 2);*/
 		height: 0;
 		flex-grow: 1;
 		writing-mode: vertical-rl;
